refactor-loginViewController: Extract header and credentials into its own views
